Posting Photos
Completed the api method for posting photos, as well as added in a
dialog to get a caption for a photo after selecting one. Also Made a
minor tweak to the selectedListModel getter to use a default if it's
null, and there's one available. This will prevent having the list tab
open and no list selected.