Live tile and list item editing
Remove live tile updating - leave this entirely to the periodic update
server.

Change the way editing list items works slightly. Instead of refreshing
the whole list when an item is added, add a blank item, get the new item
after the create call to the api, and update the blank item.